Strip padding from base64 encoded login arg param
Fixes issues when the login arg param, when encoded, ends with an
equals sign. The argument is automatically urlencoded by requests,
which would be fine if Arris's firmware could handle something as
simple as standard urlencoding, but it can't. Luckily, their
base64 decoder is tolerant of encoded strings without the padding.